Update to eslint 9
