+++ .prettierrc.yaml
@@ -1,1 +1,1 @@
-'./prettier'
+"./prettier/index.mjs"
