+++ prettier/index.mjs
@@ -1,0 +1,10 @@
+export default {
+  endOfLine: 'auto',
+  printWidth: 120,
+  singleQuote: true,
+  trailingComma: 'all',
+  plugins: [
+    'prettier-plugin-curly',
+    'prettier-plugin-packagejson',
+  ],
+};
